支持下载文件

+++ src/configs.lua
@@ -18,17 +18,14 @@ function configs.load(name)
     local path2 = path .. ".flz"
     local path3 = path .. ".mz"
 
-    local zip -- 压缩引擎
+    local compressed = false -- 压缩引擎
 
     -- 找不到原始文件，则找压缩文件
     if io.exists(path) then
         -- 找到了未压缩的文件
     elseif fastlz and io.exists(path2) then
-        zip = fastlz
+        compressed = true
         path = path2
-        -- elseif miniz and io.exists(path3) then
-        --     zip = miniz
-        --     path = path3
     else
         return false
     end
@@ -43,8 +40,8 @@ function configs.load(name)
     local data = io.readFile(path)
 
     -- 解压
-    if zip then
-        data = zip.uncompress(data)
+    if compressed then
+        data = fastlz.uncompress(data, 32 * 1024) -- 最大32KB
     end
 
     local obj, ret, err = json.decode(data)
@@ -67,21 +64,15 @@ function configs.save(name, data)
 
     -- 找文件
     local path = "/" .. name .. ".json"
-    local zip -- 压缩引擎
+    local compressed -- 压缩引擎
 
     os.remove(path)
 
     -- 大于一个block-size（flash 4k）
-    if #data > 4096 then
-        if fastlz then
-            path = path .. ".flz"
-            zip = fastlz
-            os.remove(path)
-            -- elseif miniz then
-            --     path = path .. ".mz"
-            --     zip = miniz
-            --     os.remove(path)
-        end
+    if fastlz and #data > 4096 then
+        path = path .. ".flz"
+        compressed = true
+        os.remove(path)
     end
 
     -- 删除历史(到底需不需要)，另外，是否需要备份
@@ -90,11 +81,25 @@ function configs.save(name, data)
     -- end
 
     -- 压缩
-    if zip then
-        data = zip.compress(data)
+    if compressed then
+        data = fastlz.compress(data)
     end
 
     return io.writeFile(path, data)
+end
+
+---下载配置文件，自动编码json
+---@param name string 文件名，不带.json后缀
+---@param url string 从http服务器下载
+---@return boolean 成功与否
+function configs.download(name, url)
+    log.info(tag, "download", url)
+    local code, headers, body = http.request("GET", url).wait()
+    log.info(tag, "download result", code, body)
+    -- 阻塞执行的
+    if code == 200 then
+        configs.save(name, body)
+    end
 end
 
 return configs
